+++ LaboraGTv0.1/LaboraGTv0.1/Models/Departamento.cs
@@ -23,7 +23,7 @@ namespace LaboraGTv0._1.Models
         }
     
         public string idPais { get; set; }
-        public decimal idDepartamento { get; set; }
+        public int idDepartamento { get; set; }
         public string descripcion { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
