+++ LaboraGTv0.1/LaboraGTv0.1/Models/Municipio.cs
@@ -22,8 +22,8 @@ namespace LaboraGTv0._1.Models
         }
     
         public string idPais { get; set; }
-        public decimal idDepartamento { get; set; }
-        public decimal idMunicipio { get; set; }
+        public int idDepartamento { get; set; }
+        public int idMunicipio { get; set; }
         public string descripcion { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
